Cambios sustanciosos

+++ app/Repositories/UserRepository.php
@@ -25,6 +25,17 @@ class UserRepository implements IRepository
     {
         return User::find($id);        
     }
+    public function ActivarSesion($id, $sesion)
+    {
+        $user=User::find($id);
+        $user->sesion=$sesion;
+        $user->save();
+    }
+    public function UsersAdmin()
+    {
+        $users=User::where('role_id',2)->get();
+        return $users;
+    }
     public function Update($id, $request)
     {
         $user=User::find($id);
@@ -34,6 +45,11 @@ class UserRepository implements IRepository
     public function Delete($id)
     {
         $user=User::find($id);
+        if($user->sesion==1)
+        {
+            $user =null;
+           return back()->withErrors('Esta sesion se encuemtra activa');
+        }       
         $user->delete();        
     }
 }
